requirements.txt
aiosqlite==0.21.0
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
Django==5.2.1
djangorestframework==3.16.0
dotenv==0.9.9
fixtures==4.2.5
idna==3.10
mysql-connector-python==9.3.0
parameterized==0.9.0
pycodestyle==2.5.0
python-dotenv==1.1.0
requests==2.32.3
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.13.2
urllib3==2.4.0
